Services / Home Gallery - Fixed Active button
I did a quick fix for the button toggle class, but it doesn't make sense as I was not able to simply target it using ID's as I would normally.

I suspect that the Home Gallery is because the button is not created before the toggle happens even though I can capture the buttons in an array.

Makes no sense. Now it works.